Feat: add Data source for self-hosted-agent helm values
