+++ client/agent_project_assignment_test.go
@@ -136,67 +136,4 @@ var _ = Describe("Agent Project Assignment", func() {
 			})
 		})
 	})
-
-	Describe("Agents", func() {
-		mockAgent := Agent{
-			AgentKey: "key",
-		}
-
-		expectedResponse := []Agent{mockAgent}
-
-		Describe("Successful", func() {
-			var actualResult []Agent
-			var err error
-
-			BeforeEach(func() {
-				mockOrganizationIdCall(organizationId)
-
-				httpCall = mockHttpClient.EXPECT().
-					Get("/agents", gomock.Any(), gomock.Any()).
-					Do(func(path string, request interface{}, response *[]Agent) {
-						*response = expectedResponse
-					})
-				actualResult, err = apiClient.Agents()
-			})
-
-			It("Should get organization id", func() {
-				organizationIdCall.Times(1)
-			})
-
-			It("Should send GET request with params", func() {
-				httpCall.Times(1)
-			})
-
-			It("Should return the GET result", func() {
-				Expect(actualResult).To(Equal(expectedResponse))
-			})
-
-			It("Should not return error", func() {
-				Expect(err).To(BeNil())
-			})
-		})
-
-		Describe("Failure", func() {
-			var actualResult []Agent
-			var err error
-
-			BeforeEach(func() {
-				mockOrganizationIdCall(organizationId)
-
-				httpCall = mockHttpClient.EXPECT().
-					Get("/agents", gomock.Any(), gomock.Any()).
-					Return(errorMock)
-
-				actualResult, err = apiClient.Agents()
-			})
-
-			It("Should fail if API call fails", func() {
-				Expect(err).To(Equal(errorMock))
-			})
-
-			It("Should not return results", func() {
-				Expect(actualResult).To(BeNil())
-			})
-		})
-	})
 })
